+++ libs/imports/src/Imports.hs
@@ -24,7 +24,6 @@ module Imports
     , module Data.Tuple
     , module Data.String
     , module Data.List
-    , module Debug.Trace
     , Generic
     , Typeable
     , HasCallStack
@@ -71,8 +70,6 @@ module Imports
     , unlessM
     ) where
 
-import Orphans ()
-
 -- Explicitly saying what to import because some things from Prelude clash
 -- with e.g. UnliftIO modules
 import Prelude (
@@ -100,8 +97,8 @@ import Data.Void
 import Data.Bool
 import Data.Char
 import Data.Ord
-import Data.Semigroup (Semigroup)
-import Data.Monoid
+import Data.Semigroup hiding (diff, Option, option) -- conflicts with Options.Applicative.Option (should we care?)
+import Data.Monoid hiding (First(..), Last(..)) -- First and Last are going to be deprecated. Use Semigroup instead
 import Data.Maybe
 import Data.Either
 import Data.Foldable
@@ -112,7 +109,6 @@ import Data.List hiding (insert, delete)  -- 'insert' and 'delete' are
 import Data.String
 import Control.Monad hiding (mapM_, sequence_, forM_, msum, mapM, sequence, forM)
 import Control.Monad.Extra (whenM, unlessM)
-import Debug.Trace
 import GHC.Generics (Generic)
 import Data.Typeable (Typeable)
 import GHC.Stack (HasCallStack)