+++ services/brig/src/Brig/API/Types.hs
@@ -112,6 +112,7 @@ data LoginError
     | LoginEphemeral
     | LoginPendingActivation
     | LoginThrottled RetryAfter
+    | LoginBlocked RetryAfter
 
 data ChangePasswordError
     = InvalidCurrentPassword