+++ services/brig/test/integration/API/User/Auth.hs
@@ -1,3 +1,5 @@
+{-# OPTIONS_GHC -Wno-incomplete-patterns #-}
+
 module API.User.Auth (tests) where
 
 import Imports
@@ -9,6 +11,7 @@ import Brig.Types.User.Auth
 import Brig.ZAuth (ZAuth, runZAuth)
 import UnliftIO.Async hiding (wait)
 import Control.Lens ((^?), set)
+import Control.Retry
 import Data.Aeson
 import Data.Aeson.Lens
 import Data.ByteString.Conversion
@@ -31,8 +34,9 @@ import qualified Data.Text.Lazy       as Lazy
 import qualified Brig.ZAuth           as ZAuth
 import qualified Data.Text            as Text
 import qualified Data.UUID.V4         as UUID
-
+import qualified Test.Tasty.HUnit     as HUnit
 import qualified Network.Wai.Utilities.Error as Error
+
 
 tests :: Maybe Opts.Opts -> Manager -> ZAuth.Env -> Brig -> TestTree
 tests conf m z b = testGroup "auth"
@@ -44,6 +48,7 @@ tests conf m z b = testGroup "auth"
             , test m "send-phone-code" (testSendLoginCode b)
             , test m "failure" (testLoginFailure b)
             , test m "throttle" (testThrottleLogins conf b)
+            , test m "limit-retry" (testLimitRetries conf b)
             , testGroup "sso-login"
                 [ test m "email" (testEmailSsoLogin b)
                 , test m "failure-suspended" (testSuspendedSsoLogin b)
@@ -57,6 +62,7 @@ tests conf m z b = testGroup "auth"
             , test m "unknown-cookie" (testUnknownCookie z b)
             , test m "new-persistent-cookie" (testNewPersistentCookie conf b)
             , test m "new-session-cookie" (testNewSessionCookie conf b)
+            , test m "suspend-inactive" (testSuspendInactiveUsers conf b)
             ]
         , testGroup "cookies"
             [ test m "list" (testListCookies b)
@@ -227,7 +233,54 @@ testThrottleLogins conf b = do
     liftIO $ do
         assertBool "throttle delay" (n > 0)
         threadDelay (1000000 * (n + 1))
-    void $ login b (defEmailLogin e) SessionCookie
+    login b (defEmailLogin e) SessionCookie !!! const 200 === statusCode
+
+testLimitRetries :: HasCallStack => Maybe Opts.Opts -> Brig -> Http ()
+testLimitRetries (Just conf) brig = do
+    let Just opts = Opts.setLimitFailedLogins . Opts.optSettings $ conf
+    unless (Opts.timeout opts <= 30) $
+        error "`loginRetryTimeout` is the number of seconds this test is running.  Please pick a value < 30."
+
+    usr <- randomUser brig
+    let Just email = userEmail usr
+
+    usr' <- randomUser brig
+    let Just email' = userEmail usr'
+
+    -- Login 5 times with bad password.
+    forM_ [1..Opts.retryLimit opts] $ \_ ->
+        login brig (emailLogin email defWrongPassword (Just defCookieLabel)) SessionCookie
+            <!! const 403 === statusCode
+
+    -- Login once more. This should fail for usr, even though password is correct...
+    resp <- login brig (defEmailLogin email) SessionCookie
+        <!! const 403 === statusCode
+    -- ...  but not for usr'!
+    login brig (defEmailLogin email') SessionCookie
+        !!! const 200 === statusCode
+
+    -- After the amount of time specified in "Retry-After", though,
+    -- throttling should stop and login should work again
+    do  let Just retryAfterSecs = fromByteString =<< getHeader "Retry-After" resp
+            retryTimeout = Opts.Timeout $ fromIntegral retryAfterSecs
+        liftIO $ do
+            assertBool ("throttle delay (1): " <> show (retryTimeout, Opts.timeout opts))
+                -- (this accounts for slow CI systems that lose up to 2 secs)
+                (retryTimeout >= Opts.timeout opts - 2 &&
+                 retryTimeout <= Opts.timeout opts)
+            threadDelay (1000000 * (retryAfterSecs - 2))  -- wait almost long enough.
+
+    -- fail again later into the block time window
+    rsp <- login brig (defEmailLogin email) SessionCookie <!! const 403 === statusCode
+    do  let Just retryAfterSecs = fromByteString =<< getHeader "Retry-After" rsp
+        liftIO $ do
+            assertBool ("throttle delay (2): " <> show retryAfterSecs) (retryAfterSecs <= 2)
+            threadDelay (1000000 * (retryAfterSecs + 1))  -- wait one more second, just to be safe.
+
+    -- wait long enough and login successfully!
+    liftIO $ threadDelay (1000000 * 2)
+    login brig (defEmailLogin email) SessionCookie !!! const 200 === statusCode
+
 
 -------------------------------------------------------------------------------
 -- Sso login
@@ -389,6 +442,57 @@ testNewSessionCookie config b = do
         const 200     === statusCode
         const Nothing === getHeader "Set-Cookie"
 
+testSuspendInactiveUsers :: HasCallStack => Maybe Opts.Opts -> Brig -> Http ()
+testSuspendInactiveUsers (Just config) brig = do
+    -- (context information: cookies are stored by user, not be device; so if there if the
+    -- cookie is old it means none of the devices of a user has used it for a request.)
+
+    let Just suspendAge = Opts.suspendTimeout <$> Opts.setSuspendInactiveUsers (Opts.optSettings config)
+    unless (suspendAge <= 30) $
+        error "`suspendCookiesOlderThanSecs` is the number of seconds this test is running.  Please pick a value < 30."
+
+    let check :: HasCallStack => CookieType -> String -> Http ()
+        check cookieType endPoint = do
+            user <- randomUser brig
+            let Just email = userEmail user
+            rs <- login brig (emailLogin email defPassword Nothing) cookieType
+                <!! const 200 === statusCode
+            let cky = decodeCookie rs
+
+            -- wait slightly longer than required for being marked as inactive.
+            let waitTime :: Int = floor (Opts.timeoutDiff suspendAge) + 5  -- adding 1 *should* be enough, but it's not.
+            liftIO $ threadDelay (1000000 * waitTime)
+
+            case endPoint of
+                "/access" -> do
+                    post (brig . path "/access" . cookie cky) !!! do
+                        const 403     === statusCode
+                        const Nothing === getHeader "Set-Cookie"
+                "/login" -> do
+                    login brig (emailLogin email defPassword Nothing) cookieType !!! do
+                        const 403 === statusCode
+                        const Nothing === getHeader "Set-Cookie"
+
+            let assertStatus want = do
+                  have <- retrying (exponentialBackoff 200000 <> limitRetries 6)
+                    (\_ have -> pure $ have == Suspended)
+                    (\_ -> getStatus brig (userId user))
+                  let errmsg = "testSuspendInactiveUsers: " <> show (want, cookieType, endPoint, waitTime, suspendAge)
+                  liftIO $ HUnit.assertEqual errmsg want have
+
+            assertStatus Suspended
+            setStatus brig (userId user) Active
+            assertStatus Active
+
+            login brig (emailLogin email defPassword Nothing) cookieType
+                !!! const 200 === statusCode
+
+    check SessionCookie "/access"
+    check SessionCookie "/login"
+    check PersistentCookie "/access"
+    check PersistentCookie "/login"
+
+
 -------------------------------------------------------------------------------
 -- Cookie Management
 
@@ -488,6 +592,9 @@ testTooManyCookies config b = do
                 let Just n = fromByteString =<< getHeader "Retry-After" x
                 liftIO $ threadDelay (1000000 * (n + 1))
                 loginWhenAllowed pwl t
+            403 -> error ("forbidden; " <>
+                          "perhaps setSuspendInactiveUsers.suspendTimeout is too small? " <>
+                          "(try 29 seconds).")
             xxx -> error ("Unexpected status code when logging in: " ++ show xxx)
 
 testLogout :: Brig -> Http ()
@@ -581,14 +688,6 @@ assertSaneAccessToken
 assertSaneAccessToken now uid tk = do
     assertEqual "user" uid (ZAuth.accessTokenOf tk)
     assertBool "expiry" (ZAuth.tokenExpiresUTC tk > now)
-
--- | Set user's status to something (e.g. 'Suspended').
-setStatus :: Brig -> UserId -> AccountStatus -> HttpT IO ()
-setStatus brig u s =
-    let js = RequestBodyLBS . encode $ AccountStatusUpdate s
-    in put ( brig . paths ["i", "users", toByteString' u, "status"]
-           . contentJson . Http.body js
-           ) !!! const 200 === statusCode
 
 -- | Get error label from the response (for use in assertions).
 errorLabel :: Response (Maybe Lazy.ByteString) -> Maybe Lazy.Text